Rewrite with inerhitance
